systemtests: don't fork daemons during systemtest
When ctest cancels a test (because of an error or a timeout) it will
recursively kill the whole process hierarchy. However, this does only
work if the process hierarchy is intact (i.e. if no double-fork has
happened).
To make this work correctly, we now start the daemons in the foreground
without forking which allows ctest to correctly kill them if required.